+++ src/payroll/hole10/NullTaxCalculator.java
@@ -2,7 +2,7 @@ package payroll.hole10;
 
 public class NullTaxCalculator implements TaxCalculator {
 	@Override
-	public double taxFor(double grossSalary) {
+	public long taxFor(long grossSalary) {
 		return 0;
 	}
 }
